[Refactor] Уточнение System slice
